Add bindir option to gem module
While user_install allows gem to install gem binaries into a gem
cache located in a users directory, that directory is usually not in
PATH as it's tied to a specific ruby version. In these situations,
it might be more desirable to install the binary into a more popular
location, like ~/bin.

This patch adds a bindir option to allow gem to install binaries to a
specific location.

Issue #2911